refactor : game 폴더로 폴더 구조화
